capebus: Add beaglebone board support
Introduce beaglebone capebus board support.

This patch creates the beaglebone's board cape bus controller.

The board controller is responsible for the probing of capes
at the well defined I2C address for capes, parsing the EEPROM
info and matching them to specific cape drivers.

On top of that, adapter DT enabled devices are created for
am33xx devices that have no DT bindings yet, as well as generic
devices that can be used as building blocks for the cape drivers.

Signed-off-by: Pantelis Antoniou <panto@antoniou-consulting.com>