fb: Rework locking to fix lock ordering on takeover
[The fb maintainer appears to be absent at the moment].

This is needed to fix a pile of lockdep splats that now show up because console_lock()
is being properly audited. Hugh Dickins and Sasha Levin have tested it and both reports
all looks good. This is probably not the whole story - the entire fb layer has locking
confusion problems that were previously hidden but it seems to get the ones people hit
in testing. This hopefully explains a few of the weird fb hangs that have been floating
around forever.

From: Alan Cox <alan@linux.intel.com>

Adjust the console layer to allow a take over call where the caller already
holds the locks. Make the fb layer lock in order.

This s partly a band aid, the fb layer is terminally confused about the
locking rules it uses for its notifiers it seems.

Signed-off-by: Alan Cox <alan@linux.intel.com>
Tested-by: Borislav Petkov <bp@alien8.de>