not-capebus: Beaglebone cape loader based on DT overlays.
Introduce a cape loader that is not a bus! Really!

The capes are defined in drivers/misc/cape/beaglebone/dts

Utilize the baseboard EEPROM in order to allow
per-board compatible capes and auto-loading of virtual-capes
that are soldered on the board.

This way we can support either board with the same kernel
and the same dtb.

Runtime loading of capes is also supported. Echoing the part number
to the slots sysfs node you get it loaded.

Signed-off-by: Pantelis Antoniou <panto@antoniou-consulting.com>