iio: ti_am335x_adc: Added iio_voltageX_scale
The bone-helper driver tried to display the voltage in the range 1800mV
which represents the actual ADC range of the BBB. This feature is available
in the IIO system. in_voltageX_raw points to unscaled raw values
which give the output of the ADC register directly.

in_voltageX_scale is supposed to give scaled voltages. This was
missing in the TI driver and has been added.

Signed-off-by: Zubair Lutfullah  <zubair.lutfullah@gmail.com>